refactor and make chromosome selection algorithm more liberal
The criteria for determining whether a chromosome's data should be
loaded (i.e. whether it has been 'selected') used to operate only on
strings, using the special value None to indicate that all chromosomes
should be loaded. This logic is now used to construct a predicate
function determining whether to load the chromosome. This predicate
function also works on integers.