improve fancyplotting and refactor it into a library
fancyplotting.py simply acts to load files as given on the command line,
and has the same command line interface as it did before. Instead of
doing all the calculations and plotting itself however, it calls out to
the FancyPlot library that exposes an object-oriented API for easily
creating tract length vs tract count plots.