add script to run several validation analyses on a population
Specifically, this script performs 4 inferences
 * pp classical
 * pp composite
 * pp_px classical
 * pp_px composite

The result of each analysis is plotted using FancyPlot and recorded to
files. The recorded data includes the log-likelihoods of each model as
well as the bin information and tract counts per bin in the input data.

This validation script uses the brute family of optimizers, and also
exports its findings as json