implement composite demographic model and population splitting
The population class has been augmented with a split_by_props method to
split it according to the ancestry proportions of its individuals.
Leveraging this method, the get_global_tractlengths method has been
given a keyword argument in case splitting is desired.

The log-likelihood of the composite demographic model is calculated
after combining the bins, rather than simply as the sum of the
log-likelihoods of the component models.